feat(#8): updated structure
- fixed type references
