--- src/astronomicalObject/celestial/celestial.ts
@@ -1,7 +1,0 @@
-import { AstronomicalObject } from "@/astronomicalObject/astronomicalObject";
-
-export class Celestial extends AstronomicalObject {
-  constructor() {
-    super("celestial");
-  }
-}
