+++ src/astronomicalObject/celestial/index.ts
@@ -1,1 +1,9 @@
-export { Celestial } from "./celestial";
+import { AstronomicalObject } from "@/astronomicalObject";
+
+export class Celestial extends AstronomicalObject {
+  constructor() {
+    super("celestial");
+  }
+}
+
+export default Celestial;
