+++ src/astronomicalObject/celestial/stars/index.ts
@@ -1,1 +1,9 @@
-export { Stars } from "./stars";
+import { Celestial } from "@/astronomicalObject/celestial";
+
+export class Stars extends Celestial {
+  constructor() {
+    super();
+  }
+}
+
+export default Stars;
