--- src/astronomicalObject/celestial/stars/stars.ts
@@ -1,7 +1,0 @@
-import { Celestial } from "@/astronomicalObject/celestial/celestial";
-
-export class Stars extends Celestial {
-  constructor() {
-    super();
-  }
-}
