+++ src/astronomicalObject/moon/index.ts
@@ -1,1 +1,9 @@
-export { Moon } from "./moon";
+import { AstronomicalObject } from "@/astronomicalObject";
+
+export class Moon extends AstronomicalObject {
+  constructor() {
+    super("moon");
+  }
+}
+
+export default Moon;
