--- src/astronomicalObject/planet/earth/earth.ts
@@ -1,7 +1,0 @@
-import { Planet } from "@/astronomicalObject/planet/planet";
-
-export class Earth extends Planet {
-  constructor() {
-    super("earth");
-  }
-}
