--- src/astronomicalObject/planet/planet.ts
@@ -1,11 +1,0 @@
-import { AstronomicalObject } from "@/astronomicalObject";
-
-export class Planet extends AstronomicalObject {
-  planetName: string;
-
-  constructor(name: string = "") {
-    super("planet");
-
-    this.planetName = name;
-  }
-}
