+++ src/astronomicalObject/sun/index.ts
@@ -1,2 +1,20 @@
-export { Sun } from "./sun";
-export { SunTimes } from "./times";
+import AstronomicalObject from "@/astronomicalObject";
+import { ISun, ISunProps } from "@/astronomicalObject/sun/types";
+
+export class Sun extends AstronomicalObject implements ISun {
+  private sunTime: Date;
+
+  constructor(props?: ISunProps) {
+    super("sun", props);
+
+    this.sunTime = new Date(
+      Date.UTC(
+        this.time.getUTCFullYear(),
+        this.time.getUTCMonth(),
+        this.time.getUTCDate(),
+      ),
+    );
+  }
+}
+
+export default Sun;
