--- src/location/location.ts
@@ -1,1 +1,0 @@
-export class Location {}
